Move methods from ImageController to AdsController
