--- src/main/java/ru/skypro/homework/controller/ImageApiController.java
@@ -1,61 +1,0 @@
-package ru.skypro.homework.controller;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.ArraySchema;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.util.Pair;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-import ru.skypro.homework.service.impl.ImageServiceImpl;
-
-@Slf4j
-@RestController
-@CrossOrigin(value = "http://localhost:3000")
-public class ImageApiController{
-
-    private final ImageServiceImpl imageService;
-
-    public ImageApiController(ImageServiceImpl imageService) {
-        this.imageService = imageService;
-    }
-
-    @Operation(summary = "updateAdsImage", description = "", tags = {"Изображения"})
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "OK", content = @Content(
-                    mediaType = MediaType.APPLICATION_OCTET_STREAM_VALUE,
-                    array = @ArraySchema(schema = @Schema(implementation = byte[].class)))),
-            @ApiResponse(responseCode = "404", description = "Not Found")})
-    @PatchMapping(value = "{idAds}",
-            produces = {MediaType.APPLICATION_OCTET_STREAM_VALUE},
-            consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
-    public ResponseEntity<byte[]> updateImage(Integer id, MultipartFile image) {
-        Pair<byte[], String> pair = imageService.updateImage(id, image);
-        return read(pair);
-    }
-
-    @Operation(summary = "getImage", description = "", tags = {"Изображения"})
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "OK",
-                    content = @Content(mediaType = MediaType.APPLICATION_OCTET_STREAM_VALUE,
-                            array = @ArraySchema(schema = @Schema(implementation = byte[].class)))),
-            @ApiResponse(responseCode = "404", description = "Not Found")})
-    @GetMapping(value = "{idImage}",
-            produces = {MediaType.APPLICATION_OCTET_STREAM_VALUE})
-    public ResponseEntity<byte[]> getImage(Integer idImage) {
-        Pair<byte[], String> pair = imageService.getImage(idImage);
-        return read(pair);
-    }
-
-    private ResponseEntity<byte[]> read(Pair<byte[], String> pair) {
-        return ResponseEntity.ok()
-                .contentLength(pair.getFirst().length)
-                .contentType(MediaType.parseMediaType(pair.getSecond()))
-                .body(pair.getFirst());
-    }
-}
